refactor: add menu builder builder
